Add fonts, add beta theme, clean up beta folder
