+++ packages/asc-ui/src/beta/components/shared/constants.ts
@@ -1,26 +1,2 @@
-export const HEADER_HEIGHT_SMALL = 50
-export const HEADER_HEIGHT_TALL = 158
+// eslint-disable-next-line import/prefer-default-export
 export const BACKDROP_Z_INDEX = 20
-export const TOGGLE_BUTTON_SIZE = 50
-export const INPUT_SIZE = 44
-export const BUTTON_SIZE = INPUT_SIZE
-export const SEARCH_BAR_HEIGHT = 36
-
-export const weekDays = ['ma', 'di', 'woe', 'do', 'vrij', 'za', 'zo']
-
-export const months = [
-  'januari',
-  'februari',
-  'maart',
-  'april',
-  'mei',
-  'juni',
-  'juli',
-  'augustus',
-  'september',
-  'oktober',
-  'november',
-  'december',
-]
-
-export const daysInMonth = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
