--- packages/asc-ui/src/beta/components/shared/types.ts
@@ -1,8 +1,0 @@
-export interface LoadingInterface {
-  loading?: boolean
-}
-
-export interface PageLayoutInterface {
-  tall?: boolean
-  fullWidth?: boolean
-}
