+++ packages/asc-ui/src/beta/index.ts
@@ -1,4 +1,5 @@
+import { ascBetaTheme } from './theme'
 import Button from './components/Button/Button'
 
 // eslint-disable-next-line import/prefer-default-export
-export { Button }
+export { ascBetaTheme, Button }
