+++ packages/asc-ui/src/beta/theme/default/index.ts
@@ -3,5 +3,4 @@ export { default as breakpoints } from './breakpoints'
 export { default as colors } from './colors'
 export { default as globalStyle } from './globalStyle'
 export { default as typography } from './typography'
-export { default as layouts, maxGridWidth, maxContainerWidth } from './layouts'
 export { default as spacing } from './spacing'
