+++ packages/asc-ui/src/beta/theme/index.ts
@@ -1,5 +1,4 @@
-import * as ascDefaultTheme from './default'
+import * as ascBetaTheme from './default'
 
-export { default as CompactThemeProvider } from './CompactThemeProvider'
 export { default as ThemeProvider } from './ThemeProvider'
-export { ascDefaultTheme }
+export { ascBetaTheme }
