+++ packages/asc-ui/src/beta/utils/index.ts
@@ -1,27 +1,17 @@
-import { getValueFromTheme } from './themeUtils'
 import deprecatedWarning from './deprecatedWarning'
 
 export {
   themeColor,
-  getTypographyFromTheme,
   srOnlyStyle,
   breakpoint,
   svgFill,
-  showHide,
-  ShowHideTypes,
+  fluidTypoStyle,
+  typographyStyle,
   showAboveBackDrop,
-  customCss,
-  CustomCssPropsType,
-  CustomCssPropsInterface,
   getValueFromTheme,
   themeSpacing,
 } from './themeUtils'
 export { valueFromObject } from './valueFromObject'
 export { fromProps } from './fromProps'
 
-/**
- * @deprecated Use getValueFromTheme instead
- */
-const fromTheme = getValueFromTheme
-
-export { fromTheme, deprecatedWarning }
+export { deprecatedWarning }
