+++ stories/.storybook/preview-head.html
@@ -1,3 +1,4 @@
 <!-- This is the tracking code for the fastfonts.net typography -->
 <script src="fonts/mtiFontTrackingCode.min.js"></script>
 <link href="fonts/fonts.css" rel="stylesheet">
+<link href="fonts/beta-fonts.css" rel="stylesheet">
